// src/app/interfaces/uniformes.ts
export interface Uniforme {
    PKUniformes: number;
    fkUsuario: number;
    Camisas: number;
    Pantalones: number;
    Sombreros: number;
    Botas: number;
    Chalecos: number;
    Chamarras: number;
    Impermeables: number;
    Overoles: number;
    Observaciones: string;
    FechaLlenado: Date;
    FechaActualizado: Date;
  }
  